Convert status index
